+++ NEWS.md
@@ -1,3 +1,19 @@
+# L1centrality 0.1.1.9000
+
+## Major updates
+
+* A new function `reduce_graph()` constructs the group reduced graph, which is used in computing the group L1 centrality and group L1 prestige.
+
+* A new function `L1centGROUP()` computes the group L1 centrality and group L1 prestige.  It supports undirected and directed graphs, as well as vertex and edge weighted graphs.
+
+## Minor updates
+
+* Error message of `validate_matrix()` corrected.
+
+* DESCRIPTION and the documentation `L1centrality-package` updated to include information on the group L1 centrality/prestige.
+
+* Minor typo fix in the document.
+
 # L1centrality 0.1.1
 
 ## Major updates